requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
torchinfo
torchviz


# --------- python --------- #
pandas
tqdm
pyyaml
ipywidgets
notebook
jupyterlab
matplotlib
opencv-contrib-python>=4.7.0.72
xmltodict
imageio

# --------- loggers --------- #
wandb
graphviz
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
# nvitop          # for monitoring GPU usage
# asitop          # for macOS monitoring GPU usage

